requirements.txt
pandas
sklearn
torch
transformers
nltk
pyspellchecker
